Significant reworking
Added Kotlin as a supported language for configuration.

Much improved documentation.

The creation of the *Config classes now generally use Java Futures to give the Project as chance to finish evaluating before activating some of the changes.

DefaultsExtension now supports project inheritance, getting its values from its parent project unless a value is overridden in the child.

Changed "id" to "orgId" in DefaultsExtension to make it clearer what it is for.